Moved page components into own folder
