+++ src/components/pages/about/About.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
-import img from '../../../public/img/about/IMG_1465.jpg';
 
+import img from '../../../../public/img/about/IMG_1465.jpg';
 import styles from './About.module.css';
 
 export const altText = 'Image of Lauren as a child washing a truck with dad';
