+++ tests/components/pages/projects/Projects.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
-import Projects from '../../../src/components/projects/Projects';
-import ProjectModel from '../../../src/models/projects/project';
+import Projects from '../../../../src/components/pages/projects/Projects';
+import ProjectModel from '../../../../src/models/projects/project';
 
 describe('Projects component', () => {
   const technologies = ['tech 1', 'tech 2'];
